Flesh out README.md
Flesh out README.md 

Borrowed most from ioBroker / ioBroker.zigbee README.md as it too is based on zigbee-herdsman
https://github.com/ioBroker/ioBroker.zigbee/blob/master/README.md

Not sure if you think that I should also have added this text below too after "zigpy-cc allows Zigpy to interact with..." or if that should maybe be added to Zigpy README.md instead?

The devices connected to the Zigbee-network and inform the coordinator of their status and events (button presses, motion detection, temperature change). This information is reflected in the Zigpy object-states. Some Zigpy states have feedback and send commands to the zigbee-device when the value changes (switching the state of the outlet or lamp, changing the scene or the brightness of the lamp).